requirements.txt
flask>=3.0.0
waitress>=3.0.0
reportlab>=4.0.0
msal>=1.20.0
requests>=2.28.0
pystray>=0.19.0
Pillow>=10.0.0
openpyxl>=3.1.0
werkzeug>=3.0.0
